refactored HomeSearch Component
