+++ frontend/src/Components/HomePage/style.js
@@ -1,0 +1,24 @@
+const styles = theme => ({
+    container: {
+      flexGrow: 1,
+      position: 'relative'
+    },
+    suggestionsContainerOpen: {
+      position: 'absolute',
+      zIndex: 1,
+      marginTop: theme.spacing.unit,
+      left: 0,
+      right: 0
+    },
+    suggestion: {
+      display: 'block'
+    },
+    suggestionsList: {
+      margin: 0,
+      padding: 0,
+      listStyleType: 'none'
+    }
+  });
+
+  export default styles;
+
